Improving error messages in Data.Digits
When negative or zero numeric base is used, then error is throws and it
now contains name of the "user visible" function.

Still, it would be better to find a simple way how to limit base to
positive numbers greater then zero to get rid of partial functions.

Lot of SPECIALIZED pragmas for some useful (to me) numeric types.